refactor(tests): consolidate fixtures and simplify visual test data seeding (#826)
Implements modern test data management patterns to improve maintainability
and reduce complexity in visual regression testing infrastructure.

## Changes

### Phase 1: Extract Fixtures (96% size reduction)
- Created `tests/fixtures/` directory with centralized test data
- Extracted fixtures: worlds, characters, sessions, narrative segments/decisions
- Reduced data-seeder.ts from 1,500 lines → 63 lines (96% reduction)
- Eliminated duplicate data between data-seeder.ts and seed-visual-test-data.cjs

### Phase 2: Implement Factory Pattern
- Installed fishery package for type-safe test data generation
- Created factories: worldFactory, characterFactory, sessionFactory
- Added sub-factories for attributes and skills
- Enables dynamic test data generation with TypeScript safety

### Phase 3: Simplified Seeding Logic
- Created new seedTestData.ts with single unified seeding approach
- Consolidated from 3 approaches (setState, IndexedDB, localStorage) to 1
- Created custom Playwright fixtures (seededPage, emptyStatePage)
- Automatic data seeding per test for better isolation

### Phase 4: API Mocking
- Created mockApi.ts for deterministic visual tests
- Mocked all AI endpoints (narrative, choices, world/character generation)
- Prevents visual test failures from API variability
- Added helper functions for custom responses and error states

### Phase 5: Cleanup & Migration
- Updated data-seeder.ts to re-export from new fixtures (backward compatible)
- Deleted deprecated scripts/seed-visual-test-data.cjs
- Added migration guide in data-seeder.ts comments

## Benefits

**Immediate:**
- 96% reduction in data-seeder.ts (1,500 → 63 lines)
- Single source of truth for test data
- No duplicate fixture definitions
- Type-safe factory pattern

**Long-term:**
- Easier maintenance (fixtures in one place)
- Better test isolation (Playwright fixtures)
- Reduced flakiness (mocked APIs)
- Faster test authoring (factories generate variations)

## Migration Path

Existing tests continue to work via backward-compatible data-seeder.ts.
Tests can migrate incrementally to new patterns:

```typescript
// New recommended approach
import { test } from '@/tests/visual/utils/playwright-fixtures';
test('my test', async ({ seededPage }) => {
  await seededPage.goto('/page');
});
```

Related: #820, #822, #801 (typed mock factory work)
