+++ package.json
@@ -81,6 +81,7 @@
     "eslint-config-next": "^15.5.6",
     "eslint-config-prettier": "^10.1.2",
     "eslint-plugin-storybook": "^0.12.0",
+    "fishery": "^2.3.1",
     "gh-pages": "^6.3.0",
     "glob": "^11.0.2",
     "jest": "^29.7.0",
