+++ tests/fixtures/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Centralized Test Fixtures
+ *
+ * This barrel file exports all test fixtures for easy importing across the test suite.
+ *
+ * Usage:
+ * ```typescript
+ * import { SAMPLE_WORLDS, SAMPLE_CHARACTERS } from '@/tests/fixtures';
+ * ```
+ */
+
+export { SAMPLE_WORLDS } from './worlds.fixture';
+export { SAMPLE_CHARACTERS } from './characters.fixture';
+export { SAMPLE_GAME_SESSIONS, type GameSessionFixture } from './sessions.fixture';
+export { SAMPLE_NARRATIVE_SEGMENTS, SAMPLE_DECISIONS } from './narrative.fixture';
