+++ tests/visual/utils/mockApi.ts
@@ -1,0 +1,207 @@
+import { Page } from '@playwright/test';
+
+/**
+ * Mock API endpoints for deterministic visual tests
+ *
+ * This ensures visual tests don't fail due to API variability.
+ * All AI-generated content is replaced with consistent, predictable responses.
+ *
+ * Usage:
+ * ```typescript
+ * await mockApiEndpoints(page);
+ * await page.goto('/game-session');
+ * // All AI endpoints now return consistent mock data
+ * ```
+ */
+export async function mockApiEndpoints(page: Page): Promise<void> {
+  console.log('Setting up API endpoint mocks...');
+
+  // Mock narrative generation endpoint
+  await page.route('**/api/narrative/generate', async (route) => {
+    console.log('🚫 Intercepted narrative generation API call - using mock data');
+    await route.fulfill({
+      status: 200,
+      contentType: 'application/json',
+      body: JSON.stringify({
+        segment: {
+          id: 'segment-mock-1',
+          content:
+            'Rain pelts the neon-soaked streets as you crouch behind a hover-car. The building looms ahead, its security algorithms pulsing like a digital heartbeat.',
+          type: 'scene',
+          characterIds: [],
+          metadata: {
+            mood: 'tense',
+            location: 'City streets',
+          },
+        },
+      }),
+    });
+  });
+
+  // Mock choice generation endpoint
+  await page.route('**/api/narrative/choices', async (route) => {
+    console.log('🚫 Intercepted narrative choices API call - using mock choices');
+    await route.fulfill({
+      status: 200,
+      contentType: 'application/json',
+      body: JSON.stringify({
+        choices: [
+          {
+            id: 'option-1',
+            text: 'Proceed with caution',
+            alignment: 'neutral',
+            hint: 'A balanced approach',
+          },
+          {
+            id: 'option-2',
+            text: 'Rush forward boldly',
+            alignment: 'chaotic',
+            hint: 'High risk, high reward',
+          },
+          {
+            id: 'option-3',
+            text: 'Retreat and reassess',
+            alignment: 'lawful',
+            hint: 'Play it safe',
+          },
+        ],
+      }),
+    });
+  });
+
+  // Mock world generation endpoint
+  await page.route('**/api/generate-world', async (route) => {
+    console.log('Intercepted generate-world API call');
+    await route.fulfill({
+      status: 200,
+      contentType: 'application/json',
+      body: JSON.stringify({
+        name: 'Generated Test World',
+        genre: 'Science Fiction',
+        description: 'A test world generated for visual testing',
+        attributes: [],
+        skills: [],
+        settings: {
+          toneSettings: {
+            complexity: 'medium',
+            maturityLevel: 'teen',
+            pacing: 'moderate',
+            focusAreas: ['exploration'],
+            narrativeStyle: 'adventure',
+          },
+        },
+      }),
+    });
+  });
+
+  // Mock character generation endpoint
+  await page.route('**/api/generate-character', async (route) => {
+    console.log('Intercepted generate-character API call');
+    await route.fulfill({
+      status: 200,
+      contentType: 'application/json',
+      body: JSON.stringify({
+        name: 'Generated Test Character',
+        concept: 'A test character for visual testing',
+        background: 'Generated for testing purposes',
+        personality: 'Consistent and reliable',
+      }),
+    });
+  });
+
+  // Mock world suggestion endpoint
+  await page.route('**/api/world/suggest', async (route) => {
+    console.log('Intercepted world suggestion API call');
+    await route.fulfill({
+      status: 200,
+      contentType: 'application/json',
+      body: JSON.stringify({
+        suggestions: [
+          {
+            name: 'Cyberpunk Dystopia',
+            genre: 'cyberpunk',
+            description: 'A gritty future where technology and humanity collide',
+          },
+          {
+            name: 'Epic Fantasy Realm',
+            genre: 'fantasy',
+            description: 'A magical world of dragons and ancient powers',
+          },
+        ],
+      }),
+    });
+  });
+
+  // Mock ending generation endpoint
+  await page.route('**/api/narrative/ending', async (route) => {
+    console.log('Intercepted ending generation API call');
+    await route.fulfill({
+      status: 200,
+      contentType: 'application/json',
+      body: JSON.stringify({
+        ending: {
+          text: 'Your journey comes to a close as the sun sets on the horizon.',
+          tone: 'neutral',
+          achievements: [],
+        },
+      }),
+    });
+  });
+
+  // Mock image generation endpoints (return placeholder)
+  await page.route('**/api/generate-*-image', async (route) => {
+    console.log('Intercepted image generation API call');
+    await route.fulfill({
+      status: 200,
+      contentType: 'application/json',
+      body: JSON.stringify({
+        imageUrl: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mP8/x8AAwMB/awp2z0AAAAASUVORK5CYII=',
+      }),
+    });
+  });
+
+  console.log('✅ API endpoints mocked for deterministic testing');
+}
+
+/**
+ * Mock specific narrative endpoint with custom response
+ */
+export async function mockNarrativeResponse(
+  page: Page,
+  response: { content: string; type: string }
+): Promise<void> {
+  await page.route('**/api/narrative/generate', async (route) => {
+    await route.fulfill({
+      status: 200,
+      contentType: 'application/json',
+      body: JSON.stringify({
+        segment: {
+          id: 'segment-custom',
+          content: response.content,
+          type: response.type,
+          characterIds: [],
+          metadata: {},
+        },
+      }),
+    });
+  });
+}
+
+/**
+ * Mock error response for testing error states
+ */
+export async function mockApiError(
+  page: Page,
+  endpoint: string,
+  errorMessage: string = 'Server error'
+): Promise<void> {
+  await page.route(endpoint, async (route) => {
+    await route.fulfill({
+      status: 500,
+      contentType: 'application/json',
+      body: JSON.stringify({
+        error: errorMessage,
+      }),
+    });
+  });
+}
